Using size_t instead of int for array sizes and indexes

+++ src/main/include/functions.h
@@ -5,7 +5,7 @@
 /**
  * Returns the size of a triangular matrix, without the diagonal.
  */
-int size_triangular_matrix(const int n);
+size_t size_triangular_matrix(const size_t n);
 
 /**
  * Computes the overlap between two particles.
@@ -17,25 +17,25 @@ double compute_overlap(const Particle p1, const Particle p2);
  * Takes two arrays of vectors,
  * and sums together the vectors with the same index.
  */
-void sum_vectors(const int size, const Vector *v1, const Vector *v2, Vector *result);
+void sum_vectors(const size_t size, const Vector *v1, const Vector *v2, Vector *result);
 
 /**
  * Applies the forces to the particles with the same index,
  * and computes the resultant acceleration.
  */
-void compute_acceleration(const int size, const ParticleProperties *particles, const Vector *forces, Vector *resultant_accelerations);
+void compute_acceleration(const size_t size, const ParticleProperties *particles, const Vector *forces, Vector *resultant_accelerations);
 
 /**
  * Derives the resultant velocity,
  * of an initial velocity with an applied acceleration for given a time delta.
  */
-void compute_velocity(const int size, const Vector *accelerations, const double dt, Vector *velocities);
+void compute_velocity(const size_t size, const Vector *accelerations, const double dt, Vector *velocities);
 
 /**
  * Computes the displacement of the particles,
  * with an applied velocity for a given time delta.
  */
-void compute_displacement(const int size, const double dt, const Vector *velocities, Vector *displacements);
+void compute_displacement(const size_t size, const double dt, const Vector *velocities, Vector *displacements);
 
 /**
  * Computes the contacts between all particles.
@@ -43,4 +43,4 @@ void compute_displacement(const int size, const double dt, const Vector *velocit
  * Note: The size of the buffer should be equals to
  *       the size of a triangular matrix for the number of particles.
  */
-int compute_contacts(const int size, const Particle *particles, Contact *contacts_buffer);
+size_t compute_contacts(const size_t size, const Particle *particles, Contact *contacts_buffer);
